// operator-app/lib/types/database.ts
export interface ApiResponse<T> {
  data: T;
  error?: string;
  message?: string;
}